pyproject.toml
[project]
name = "langchain-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.1.0",
    "isort>=8.0.0",
    "langchain>=1.2.10",
    "langchain-core>=1.2.14",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.10",
    "python-dotenv>=1.2.1",
]
